# Client/RetroLauncher.Client/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RetroLauncher Api",
    "version": "v1"
  },
  "paths": {
    "/api/v{version}/Game/getList": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Genres": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32"
                    }
                  },
                  "Platforms": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32"
                    }
                  },
                  "Limit": {
                    "maximum": 100,
                    "minimum": 1,
                    "type": "integer",
                    "format": "int32"
                  },
                  "Offset": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "encoding": {
                "Name": {
                  "style": "form"
                },
                "Genres": {
                  "style": "form"
                },
                "Platforms": {
                  "style": "form"
                },
                "Limit": {
                  "style": "form"
                },
                "Offset": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Game.Get.GamesGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Game.Get.GamesGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Game.Get.GamesGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Game/getbyid": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Game.Get.GameGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Game.Get.GameGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Game.Get.GameGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Game/getrom": {
      "get": {
        "tags": [
          "Game"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v{version}/Genre/getList": {
      "get": {
        "tags": [
          "Genre"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Genre.Get.GengresGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Genre.Get.GengresGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Genre.Get.GengresGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Platform/getList": {
      "get": {
        "tags": [
          "Platform"
        ],
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Platform.PlatformsGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Platform.PlatformsGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.Platform.PlatformsGetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RetroLauncher.WebAPI.Controllers.v1.ErrorGetResponse": {
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Controllers.v1.Game.Get.GameGetResponse": {
        "type": "object",
        "properties": {
          "game": {
            "$ref": "#/components/schemas/RetroLauncher.WebAPI.Models.Game"
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Controllers.v1.Game.Get.GamesGetResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RetroLauncher.WebAPI.Models.Game"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Controllers.v1.Genre.Get.GengresGetResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RetroLauncher.WebAPI.Models.Genre"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Controllers.v1.Platform.PlatformsGetResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RetroLauncher.WebAPI.Models.Platform"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Models.Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "nameSecond": {
            "type": "string",
            "nullable": true
          },
          "nameOther": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "developer": {
            "type": "string",
            "nullable": true
          },
          "annotation": {
            "type": "string",
            "nullable": true
          },
          "genre": {
            "type": "string",
            "nullable": true
          },
          "platform": {
            "$ref": "#/components/schemas/RetroLauncher.WebAPI.Models.Platform"
          },
          "downloads": {
            "type": "integer",
            "format": "int32"
          },
          "ratings": {
            "type": "number",
            "format": "double"
          },
          "gameLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RetroLauncher.WebAPI.Models.GameLink"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Models.GameLink": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "nullable": true
          },
          "typeUrl": {
            "$ref": "#/components/schemas/RetroLauncher.WebAPI.Models.TypeUrl"
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Models.Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "genreName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Models.Platform": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "platformName": {
            "type": "string",
            "nullable": true
          },
          "alias": {
            "type": "string",
            "nullable": true
          },
          "urlPic": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RetroLauncher.WebAPI.Models.TypeUrl": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  }
}